Refactor: move cartStore file to api folder
