+++ week8/unique0902/package.json
@@ -45,6 +45,7 @@
     "@types/node": "^20.9.0",
     "@types/react": "^18.2.37",
     "@types/react-dom": "^18.2.15",
+    "@types/styled-components": "^5.1.30",
     "typescript": "^5.2.2"
   }
 }
